Created Postgres Utils header with common auxiliary methods
